enable useCustomAnchors
