+++ packages/typedoc-plugin/src/markdown-theme-context.ts
@@ -1,7 +1,7 @@
 import { MarkdownThemeContext } from 'typedoc-plugin-markdown'
 
 import { getReflectionFlags } from './theme/context/helpers/get-reflection-flags'
-import { memberContainer } from './theme/context/partials/member.container'
+import { memberContainerPatch } from './theme/context/partials/member.container.patch'
 import { declaration } from './theme/context/partials/member.declaration'
 import { declarationTitle } from './theme/context/partials/member.declarationTitle'
 import { inheritance } from './theme/context/partials/member.inheritance'
@@ -12,12 +12,14 @@ export class MyMarkdownThemeContext extends MarkdownThemeContext {
   constructor(...args: ConstructorParameters<typeof MarkdownThemeContext>) {
     super(...args)
 
+    const partials = { ...this.partials }
+
     this.helpers.getReflectionFlags = (...args) => {
       return getReflectionFlags.call(this, ...args)
     }
 
     this.partials.memberContainer = (...args) => {
-      return memberContainer.call(this, ...args)
+      return memberContainerPatch.call(this, ...args) || partials.memberContainer.call(this, ...args)
     }
     this.partials.inheritance = (...args) => {
       return inheritance.call(this, ...args)
